I added a few things around the performance specs.

1) The data that was being printed for the performance specs was not the same as being used for the `expect` statements. Each benchmark was being run twice, once for the `expect` and once for the printed stats.

2) I added some info to the stats message. It now looks like:
```
Serialize to Ruby Hash 1 records (speed goal: 25x)
Serializer      Records    Time       Speed Up
Fast serializer 1          0.08 ms
AMS serializer  1          0.88 ms    10.54x ✘

Serialize to JSON string 25 records (speed goal: 25x)
Serializer      Records    Time       Speed Up
Fast serializer 25         0.44 ms
AMS serializer  25         15.16 ms   34.62x ✔
```

3) I renamed things from `our_*` to `fast_jsonapi_*` to be more explicit.
4) Updated `README` with info on running/not running performance specs.