@guilleiguaran 2018-02-15
This PR should be sent to`dev` branch.