@hmcfletch 2018-02-15
closing in favor of #86